Add minimal Travis configuration and deps
Specify a version of Perl which matches current Debian Stable (Jessie).

Add dependencies to Build.PL, so Travis knows to install them (and just
for general correctness). Dpkg::* modules are commented out because they
are not on CPAN and this causes problems installing the prerequisites.

Add a badge for the integration branch to README.