Fix misdetecting loops for -dev packages
Don't consider package binaries depending on other binaries in the same
source to be a loop. This can happen quite often with -dev packages that
depend on the actual library package, but since they both get built at
once in a single build of the source, it is not a dependency loop.